kafka_scripts: Do not subscribe to ALERT_TOPIC too soon
Move the consumer creation and subscription code inside the function so
that the topic won't be subscribed to at import time.  This is important
for two reasons: (a) avoiding import time side effects and (b)
subscribing to a topic creates it as a side effect, but it will create
it with different configuration as in the initialize_services module and
will therefore cause an error when doing the initialization.