Change the datahubhel.core models
The models in the datahubhel.core needs some refactoring to
comply with the STA. Therefore add the Location model and
attach it to Thing. Also, the Sensor model no longer serves the
purpose of physical sensor. It is more like a parent umbrella
under which there are other sensors. E.g there could be several
Ta120sensor of type Sensor.