Make sure that many2many field on TA120Sensor get saved
The datastream needs to be attached to each TA120Sensor upon
creation. Since datastream is the M2M field, it threfore needs
to exist before it can be attached to TA120Sensor. This is why
we needed the transaction.commit so that the creating and attaching
of datastreams to TA120Sensor works on post_save.