Add extra information to ServiceSerializer
The service needs to display all the datastreams owned by the
user so that while subscribing to the service user can decide
whether to give the permission to read the datastream or not.
Also, add the allowed_permissions to track the list of datastreams
the user have given permission to be read.